Add files via upload
I created this DynDistCreator script because there was nothing found that was an easy way out of creating Dynamic Distribution Groups in Exchange Online using Powershell. The problem was that the standard web interface did not give a lot of possibilities to create a broader way of creating DDL in Exchange online. So I created this to help Admins make a DDL in a way that gives a better overview. Remember that attributes are not always self-explaining.